+++ app/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
         android:layout_height="wrap_content"
         android:text="@string/activity_main_saved_values_label"
         android:textColor="@color/black"
-        android:textSize="@dimen/activity_main_text_size"
+        android:textSize="@dimen/text_size_middle"
         app:layout_constraintBottom_toTopOf="@id/currentValueText"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -35,7 +35,7 @@
         android:layout_height="wrap_content"
         android:text="@string/activity_main_current_value_label"
         android:textColor="@color/black"
-        android:textSize="@dimen/activity_main_text_size"
+        android:textSize="@dimen/text_size_middle"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -46,12 +46,12 @@
         android:id="@+id/currentValueResult"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/current_value_result_start_margin"
-        android:layout_marginTop="@dimen/current_value_result_top_margin"
-        android:layout_marginBottom="@dimen/current_value_result_bottom_margin"
+        android:layout_marginStart="@dimen/margin_large"
+        android:layout_marginTop="@dimen/margin_small"
+        android:layout_marginBottom="@dimen/margin_small"
         android:gravity="center"
         android:textColor="@color/black"
-        android:textSize="@dimen/activity_main_text_size"
+        android:textSize="@dimen/text_size_middle"
         app:layout_constraintBottom_toBottomOf="@id/saveButton"
         app:layout_constraintEnd_toStartOf="@id/saveButton"
         app:layout_constraintStart_toStartOf="parent"
@@ -63,9 +63,9 @@
         style="@style/ActivityMainButtons"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/save_button_top_margin"
-        android:layout_marginEnd="@dimen/save_button_end_margin"
-        android:padding="@dimen/save_button_padding"
+        android:layout_marginTop="@dimen/margin_small"
+        android:layout_marginEnd="@dimen/margin_middle"
+        android:padding="@dimen/padding_big"
         android:text="@string/activity_main_save_button_label"
         app:layout_constraintBottom_toTopOf="@id/calculatorButton"
         app:layout_constraintEnd_toEndOf="parent"
@@ -77,8 +77,8 @@
         style="@style/ActivityMainButtons"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginHorizontal="@dimen/calculator_button_horizontal_margin"
-        android:padding="@dimen/calculator_button_padding"
+        android:layout_marginHorizontal="@dimen/margin_middle"
+        android:padding="@dimen/padding_big"
         android:text="@string/activity_main_calculator_button_label"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
