+++ app/src/main/res/values-night/themes.xml
@@ -13,31 +13,36 @@
         <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
     </style>
+
     <style name="CharButton">
         <item name="backgroundTint">@color/grey</item>
         <item name="android:textColor">@color/black</item>
         <item name="android:textStyle">bold</item>
     </style>
+
     <style name="SavedValues">
-        <item name="android:layout_marginTop">@dimen/saved_values_top_margin</item>
+        <item name="android:layout_marginTop">@dimen/margin_large</item>
         <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">@dimen/activity_main_text_size</item>
+        <item name="android:textSize">@dimen/text_size_middle</item>
     </style>
+
     <style name="SavedValuesLandscape">
-        <item name="android:layout_marginTop">@dimen/saved_values_top_margin</item>
+        <item name="android:layout_marginTop">@dimen/margin_large</item>
         <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">@dimen/activity_main_text_size_landscape</item>
+        <item name="android:textSize">@dimen/text_size_middle</item>
     </style>
+
     <style name="ActivityMainButtons">
         <item name="android:textAllCaps">false</item>
         <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">@dimen/activity_main_text_size_for_buttons</item>
+        <item name="android:textSize">@dimen/text_size_small</item>
         <item name="backgroundTint">@color/grey</item>
     </style>
+
     <style name="ActivityMainButtonsLandscape">
         <item name="android:textAllCaps">false</item>
         <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">@dimen/activity_main_text_size_for_buttons_landscape</item>
+        <item name="android:textSize">@dimen/text_size_small</item>
         <item name="backgroundTint">@color/grey</item>
     </style>
 </resources>
