+++ app/src/main/res/values/dimens.xml
@@ -1,28 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="top_margin">4dp</dimen>
-    <dimen name="start_margin">8dp</dimen>
-    <dimen name="saved_values_top_margin">32dp</dimen>
-    <dimen name="current_value_result_start_margin">32dp</dimen>
-    <dimen name="current_value_result_top_margin">8dp</dimen>
-    <dimen name="current_value_result_top_margin_landscape">16dp</dimen>
-    <dimen name="current_value_top_margin_landscape">16dp</dimen>
-    <dimen name="save_button_top_margin">8dp</dimen>
-    <dimen name="end_margin">8dp</dimen>
-    <dimen name="save_button_end_margin">12dp</dimen>
-    <dimen name="calculator_button_horizontal_margin">12dp</dimen>
-    <dimen name="calculator_button_top_margin_landscape">8dp</dimen>
-    <dimen name="current_value_result_bottom_margin">8dp</dimen>
-    <dimen name="calculator_button_bottom_margin">8dp</dimen>
+    <dimen name="margin_small">8dp</dimen>
+    <dimen name="margin_middle">12dp</dimen>
+    <dimen name="margin_big">16dp</dimen>
+    <dimen name="margin_large">32dp</dimen>
 
-    <dimen name="end_padding">4dp</dimen>
-    <dimen name="save_button_padding">12dp</dimen>
-    <dimen name="calculator_button_padding">12dp</dimen>
-    <dimen name="current_value_result_padding_end">8dp</dimen>
+    <dimen name="padding_small">4dp</dimen>
+    <dimen name="padding_middle">8dp</dimen>
+    <dimen name="padding_big">12dp</dimen>
 
-    <dimen name="shown_text_size">32sp</dimen>
-    <dimen name="activity_main_text_size">24sp</dimen>
-    <dimen name="activity_main_text_size_landscape">24sp</dimen>
-    <dimen name="activity_main_text_size_for_buttons">18sp</dimen>
-    <dimen name="activity_main_text_size_for_buttons_landscape">20sp</dimen>
+    <dimen name="text_size_small">18sp</dimen>
+    <dimen name="text_size_middle">24sp</dimen>
+    <dimen name="text_size_big">32sp</dimen>
 </resources>
